postgrep/requirements.txt
# requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
python-jose[cryptography]

passlib[bcrypt]
email-validator
